Draftmancer utils & Front Cards in DB
Moved all draftmancer stuff to utils folder. Moved stuff for fetching HCJ front cards into backend, and added HCJ front cards to the database as tokens
