+++ packages/backend/src/fetchdatabase.ts
@@ -570,7 +570,7 @@ export const fetchDatabase = async (usingApproved: boolean = false) => {
           );
         }
       } else if (!index && !face.layout) {
-        addPropToFace(cardObject, 'layout', getFrontLayout(face) || HCLayout.Front, index);
+        addPropToFace(cardObject, 'layout', getFrontLayout(face) || HCLayout.Normal, index);
       } else {
         addPropToFace(cardObject, 'layout', HCLayout.Split, index);
       }
