+++ packages/frontend/src/hells-cubes/CubeResources.tsx
@@ -6,7 +6,7 @@ import { toCockCube } from '@hellfall/shared/utils';
 import { useAtomValue } from 'jotai';
 import { HCCard } from '@hellfall/shared/types';
 import { ReactNode } from 'react';
-import { getDraftmancerForCube } from './draftmancer';
+import { downloadDraftmancer } from './downloadDraftmancer.ts';
 import { getHc5 } from './getHc5.ts';
 import { toMPCAutofill } from './toMPCAutofill.ts';
 import { getLands } from './getLands.ts';
@@ -231,11 +231,11 @@ export const CubeResources = () => {
                 <button
                   onClick={() => {
                     const val = toCockCube({
-                      set: cubeSetup.id,
                       name: cubeSetup.name,
-                      allCards: cards,
+                      set: cubeSetup.id,
                       cardList:
                         cubeSetup.id == 'HC5' ? getHc5() : getSplitSet(cards, cubeSetup.id).cards,
+                      allCards: cards,
                     });
 
                     const url = 'data:text/plain;base64,' + btoa(unescape(encodeURIComponent(val)));
@@ -257,10 +257,14 @@ export const CubeResources = () => {
                 ) : (
                   <button
                     onClick={() => {
-                      getDraftmancerForCube({
-                        id: cubeSetup.id,
+                      downloadDraftmancer({
                         name: cubeSetup.name,
+                        set: cubeSetup.id,
                         allCards: cards,
+                        cardList:
+                          cubeSetup.id == 'HC5'
+                            ? getHc5()
+                            : getSplitSet(cards, cubeSetup.id, true).cards,
                       });
                     }}
                   >
