--- packages/frontend/src/hells-cubes/draftmancer/index.ts
@@ -1,2 +1,0 @@
-export { toDraftmancerCube } from './toDraftmancer.ts';
-export { getDraftmancerForCube } from './getDraftmancerForCube.ts';
