+++ packages/shared/src/data/index.ts
@@ -1,1 +1,1 @@
-export * from '.';
+export * from './sets';
