+++ packages/shared/src/data/sets.ts
@@ -14,6 +14,8 @@ export const extraSetList = [
   'HCV.L',
   'NRM',
   'HCT',
+  'FHCJ',
+  'HBB',
   'SFT',
 ];
 export const allSetsList = [
@@ -58,6 +60,8 @@ export const allSetsList = [
   'HCV.L',
   'NRM',
   'HCT',
+  'FHCJ',
+  'HBB',
   'SFT',
   'NotMagic',
 ];
