+++ packages/shared/src/data/types.json
@@ -733,6 +733,7 @@
 		"Michael",
 		"Microtransaction",
 		"Middle",
+		"Mii",
 		"Milk",
 		"Mime",
 		"Mind",
