+++ packages/shared/src/utils/index.ts
@@ -5,5 +5,5 @@ export * from './textHandling.ts';
 export * from './pipsHandling.ts';
 export * from './cardHandling';
 export * from './cockatrice';
-// export * from './draftmancer'
+export * from './draftmancer'
 // export * from './tts'
